[TH2-5267] Split ErrorCollector to interface and implementation
