+++ src/test/java/com/exactpro/th2/mstore/TestErrorCollector.java
@@ -1,5 +1,5 @@
 /*
- * Copyright 2023 Exactpro (Exactpro Systems Limited)
+ * Copyright 2023-2025 Exactpro (Exactpro Systems Limited)
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -51,8 +51,8 @@ import static org.mockito.Mockito.verifyNoMoreInteractions;
 @SuppressWarnings("DynamicRegexReplaceableByCompiledPattern")
 class TestErrorCollector {
 
-    private static final long PERIOD = RandomUtils.nextLong(0, Long.MAX_VALUE);
-    private static final TimeUnit TIME_UNIT = TimeUnit.values()[RandomUtils.nextInt(0, TimeUnit.values().length)] ;
+    private static final long PERIOD = RandomUtils.secure().randomLong(0, Long.MAX_VALUE);
+    private static final TimeUnit TIME_UNIT = TimeUnit.values()[RandomUtils.secure().randomInt(0, TimeUnit.values().length)] ;
 
     @Mock
     private Logger logger;
@@ -75,7 +75,7 @@ class TestErrorCollector {
     @BeforeEach
     void beforeEach() {
         doReturn(future).when(executor).scheduleAtFixedRate(any(Runnable.class), anyLong(), anyLong(), any(TimeUnit.class));
-        errorCollector = new ErrorCollector(executor, eventRouter, rootEvent, PERIOD, TIME_UNIT);
+        errorCollector = EventErrorCollector.create(executor, eventRouter, rootEvent, PERIOD, TIME_UNIT);
         verify(executor).scheduleAtFixedRate(taskCaptor.capture(), eq(PERIOD), eq(PERIOD), eq(TIME_UNIT));
         verifyNoMoreInteractions(executor);
     }
